Extracted the data filling parts into DataFaker
